# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: ExpertMatch API
  version: 1.0.0
  description: |
    GraphRAG-powered expert discovery and matching service for RFPs and team formation.
    This API enables automatic client generation for frontend applications using OpenAPI generators.
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: ""
    description: Current Server (auto-detected)
  - url: http://localhost:8093
    description: Local Development Server
  - url: http://192.168.0.73:8093
    description: Remote Development Server
  - url: https://api.expertmatch.com
    description: Production Server

# Security: Authentication and authorization are handled by Spring Gateway.
# Spring Gateway validates JWT tokens and populates user information in HTTP headers:
# - X-User-Id: User identifier (required)
# - X-User-Roles: Comma-separated list of roles (optional, e.g., "ROLE_USER,ROLE_ADMIN")
# - X-User-Email: User email address (optional)
...
components:
...
  schemas:
...
    TestDataSizeResponse:
      type: object
      required:
        - success
        - message
        - size
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        message:
          type: string
          description: Operation message
        size:
          type: string
          enum: [ tiny, small, medium, large, huge ]
          description: Dataset size that was generated
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        message:
          type: string
          description: Operation message
    IngestionResult:
      type: object
      required:
        - totalProfiles
        - successCount
        - errorCount
        - results
        - sourceName
      properties:
        totalProfiles:
          type: integer
          description: Total number of profiles processed
          example: 10
        successCount:
          type: integer
          description: Number of successfully processed profiles
          example: 9
        errorCount:
          type: integer
          description: Number of profiles that failed to process
          example: 1
        results:
          type: array
          description: Detailed results for each processed profile
          items:
            $ref: '#/components/schemas/ProcessingResult'
        sourceName:
          type: string
          description: Name or path of the source (file or directory)
          example: "classpath:data"
    ProcessingResult:
      type: object
      required:
        - employeeId
        - employeeName
        - success
        - projectsProcessed
        - projectsSkipped
        - projectErrors
      properties:
        employeeId:
          type: string
          description: Employee ID
          example: "4000741400013306668"
        employeeName:
          type: string
          description: Employee name
          example: "John Doe"
        success:
          type: boolean
          description: Whether the profile was processed successfully
        errorMessage:
          type: string
          nullable: true
          description: Error message if processing failed
        projectsProcessed:
          type: integer
          description: Number of projects successfully processed
          example: 5
        projectsSkipped:
          type: integer
          description: Number of projects skipped due to errors
          example: 0
        projectErrors:
          type: array
          description: List of error messages for failed projects
          items:
            type: string
          example: [ ]

    # Error schemas
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp
        path:
          type: string
          description: API path that caused the error
        details:
          type: object
          description: Additional error details
          additionalProperties: true
paths:
...
  /ingestion/json-profiles:
    post:
      tags:
        - Ingest
      summary: Ingest expert profiles from JSON files
      description: |
        Ingests expert profiles from JSON files. Supports both array format (multiple profiles) 
        and single object format (backward compatible). Handles partial data gracefully with defaults.
        Requires ADMIN role.
      operationId: ingestJsonProfiles
      parameters:
        - name: directory
          in: query
          description: |
            Directory path containing JSON profile files. Supports classpath (e.g., "classpath:data") 
            or file system paths. If not specified, defaults to "classpath:data".
          required: false
          schema:
            type: string
          example: "classpath:data"
        - name: file
          in: query
          description: |
            Single JSON file path to ingest. Supports classpath (e.g., "classpath:data/profile.json") 
            or file system paths. If specified, directory parameter is ignored.
          required: false
          schema:
            type: string
          example: "classpath:data/siarhei-berdachuk-profile.json"
      responses:
        '200':
          description: Ingestion completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResult'
        '400':
          description: Invalid request or JSON parsing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (requires ADMIN role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error during ingestion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /test-data:
    post:
      tags:
        - Test Data
      summary: Generate test data
      description: |
        Generates synthetic test data using Datafaker. Requires ADMIN role.
        Size options: 'tiny' (5 employees), 'small' (50), 'medium' (500), 
        'large' (2,000), or 'huge' (50,000).
      operationId: generateTestData
      parameters:
        - name: size
          in: query
          required: false
          description: |
            Predefined dataset size. Options:
            - tiny: 5 employees, 5 projects, ~15 work experiences
            - small: 50 employees, 100 projects, ~250 work experiences
            - medium: 500 employees, 1,000 projects, ~4,000 work experiences
            - large: 2,000 employees, 4,000 projects, ~20,000 work experiences
            - huge: 50,000 employees, 100,000 projects, ~750,000 work experiences
          schema:
            type: string
            enum: [ tiny, small, medium, large, huge ]
            default: small
          example: small
        - name: clear
          in: query
          required: false
          description: |
            If true, clears all existing test data (employees, work experience, projects, technologies) 
            before generating new data. Default: false (new data will be appended to existing data).
          schema:
            type: boolean
            default: false
          example: false
      responses:
        '200':
          description: Test data generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestDataSizeResponse'
        '400':
          description: Invalid size parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /test-data/embeddings:
    post:
      tags:
        - Test Data
      summary: Generate embeddings
      description: |
        Generates vector embeddings for all work experience records using the configured embedding model.
        Requires ADMIN role.
      operationId: generateEmbeddings
      responses:
        '200':
          description: Embeddings generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /test-data/graph:
    post:
      tags:
        - Test Data
      summary: Build graph relationships
      description: |
        Builds graph relationships in Apache AGE from existing database data. Requires ADMIN role.
      operationId: buildGraph
      responses:
        '200':
          description: Graph relationships built successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /test-data/complete:
    post:
      tags:
        - Test Data
      summary: Generate complete dataset
      description: |
        Generates complete dataset including test data, embeddings, and graph relationships in one operation.
        Requires ADMIN role. Size options: 'tiny' (5 employees), 'small' (50), 'medium' (500), 
        'large' (2,000), or 'huge' (50,000).
      operationId: generateCompleteDataset
      parameters:
        - name: size
          in: query
          required: false
          description: |
            Predefined dataset size. Options:
            - tiny: 5 employees, 5 projects, ~15 work experiences
            - small: 50 employees, 100 projects, ~250 work experiences
            - medium: 500 employees, 1,000 projects, ~4,000 work experiences
            - large: 2,000 employees, 4,000 projects, ~20,000 work experiences
            - huge: 50,000 employees, 100,000 projects, ~750,000 work experiences
          schema:
            type: string
            enum: [ tiny, small, medium, large, huge ]
            default: small
          example: small
        - name: clear
          in: query
          required: false
          description: |
            If true, clears all existing test data (employees, work experience, projects, technologies) 
            before generating new data. Default: false (new data will be appended to existing data).
          schema:
            type: boolean
            default: false
          example: false
      responses:
        '200':
          description: Complete dataset generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestDataSizeResponse'
        '400':
          description: Invalid size parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - ADMIN role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'